# docs/swagger.yaml
basePath: /api
definitions:
  apperrors.Code:
    enum:
    - 1
    - 2
    - 3
    - 4
    - 5
    - 6
    - 7
    type: integer
    x-enum-varnames:
    - InternalCode
    - BadRequestCode
    - NotFoundCode
    - AlreadyExistsCode
    - UnauthorizedCode
    - ForbiddenCode
    - VersionMismatchCode
  apperrors.Error:
    properties:
      code:
        $ref: '#/definitions/apperrors.Code'
      incidentID:
        type: string
      stackTrace:
        type: string
    type: object
...
  contracts.CreateMovieRequest:
    properties:
      cast:
        items:
          $ref: '#/definitions/contracts.MovieCreditInfo'
        type: array
      description:
        type: string
      genreIds:
        items:
          type: integer
        type: array
      releaseDate:
        type: string
      title:
        maxLength: 100
        minLength: 1
        type: string
    type: object
  contracts.Genre:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  contracts.Movie:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        type: integer
      releaseDate:
        type: string
      title:
        type: string
    type: object
  contracts.MovieCredit:
    properties:
      details:
        type: string
      role:
        type: string
      star:
        $ref: '#/definitions/contracts.Star'
    type: object
  contracts.MovieCreditInfo:
    properties:
      details:
        type: string
      role:
        type: string
      starId:
        type: integer
    type: object
  contracts.MovieDetails:
    properties:
      cast:
        items:
          $ref: '#/definitions/contracts.MovieCredit'
        type: array
      createdAt:
        type: string
      deletedAt:
        type: string
      description:
        type: string
      genres:
        items:
          $ref: '#/definitions/contracts.Genre'
        type: array
      id:
        type: integer
      releaseDate:
        type: string
      title:
        type: string
      version:
        type: integer
    type: object
...
  contracts.PaginatedRequestOrdered:
    properties:
      order:
        type: string
      page:
        type: integer
      size:
        type: integer
      sort:
        type: string
    type: object
  contracts.Star:
    properties:
      bio:
        type: string
      birthDate:
        type: string
      birthPlace:
        type: string
      createdAt:
        type: string
      deathDate:
        type: string
      deletedAt:
        type: string
      firstName:
        type: string
      id:
        type: integer
      lastName:
        type: string
      middleName:
        type: string
    type: object
  contracts.UpdateMovieRequest:
    properties:
      cast:
        items:
          $ref: '#/definitions/contracts.MovieCreditInfo'
        type: array
      description:
        type: string
      genreIds:
        items:
          type: integer
        type: array
      releaseDate:
        type: string
      title:
        maxLength: 100
        type: string
      version:
        type: integer
    type: object
...
  genres.CreateGenreRequest:
    properties:
      name:
        maxLength: 32
        minLength: 3
        type: string
    type: object
  genres.Genre:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  genres.UpdateGenreRequest:
    properties:
      name:
        maxLength: 32
        minLength: 3
        type: string
    type: object
...
  pagination.PaginatedResponseOrdered-contracts_Movie:
    properties:
      items:
        items:
          $ref: '#/definitions/contracts.Movie'
        type: array
      order:
        type: string
      page:
        minimum: 0
        type: integer
      size:
        minimum: 0
        type: integer
      sort:
        type: string
      total:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: david417program@gmail.com
    name: API Creator Support
    url: https://github.com/DavidMovas/Movies-Reviews
  description: This API is used for managing movies reviews.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Movies Reviews API
  version: 1.0.0
paths:
...
  /genres:
    get:
      description: Get all genres
      operationId: get-genres
      produces:
      - application/json
      responses:
        "200":
          description: Genres, or nil if none found
          schema:
            items:
              $ref: '#/definitions/genres.Genre'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apperrors.Error'
      tags:
      - genres
    post:
      description: Create new genre
      operationId: create-genre
      parameters:
      - description: Genre
        in: body
        name: genre
        required: true
        schema:
          $ref: '#/definitions/genres.CreateGenreRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Genre
          schema:
            $ref: '#/definitions/genres.Genre'
        "400":
          description: Invalid parameter or missing parameter
          schema:
            $ref: '#/definitions/apperrors.Error'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/apperrors.Error'
        "409":
          description: Genre with that name already exists
          schema:
            $ref: '#/definitions/apperrors.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apperrors.Error'
      tags:
      - genres
  /genres/{genreId}:
    delete:
      description: Delete genre by id
      operationId: delete-genre-by-id
      parameters:
      - description: Genre ID
        in: path
        name: genreId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Genre deleted (softly deleting)
        "400":
          description: Invalid genre id, invalid parameter or missing parameter
          schema:
            $ref: '#/definitions/apperrors.Error'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/apperrors.Error'
        "404":
          description: Genre not found
          schema:
            $ref: '#/definitions/apperrors.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apperrors.Error'
      tags:
      - genres
    get:
      description: Get genre by id
      operationId: get-genre-by-id
      parameters:
      - description: Genre ID
        in: path
        name: genreId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Genre
          schema:
            $ref: '#/definitions/genres.Genre'
        "400":
          description: Invalid genre id, invalid parameter or missing parameter
          schema:
            $ref: '#/definitions/apperrors.Error'
        "404":
          description: Genre not found
          schema:
            $ref: '#/definitions/apperrors.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apperrors.Error'
      tags:
      - genres
    put:
      description: Update genre by id
      operationId: update-genre-by-id
      parameters:
      - description: Genre ID
        in: path
        name: genreId
        required: true
        type: integer
      - description: Genre
        in: body
        name: genre
        required: true
        schema:
          $ref: '#/definitions/genres.UpdateGenreRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Genre updated
        "400":
          description: Invalid genre id, invalid parameter or missing parameter
          schema:
            $ref: '#/definitions/apperrors.Error'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/apperrors.Error'
        "404":
          description: Genre not found
          schema:
            $ref: '#/definitions/apperrors.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apperrors.Error'
      tags:
      - genres
  /movies:
    get:
      description: Get movies
      operationId: get-movies
      parameters:
      - description: Request, if request body empty, default values will be used
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/contracts.PaginatedRequestOrdered'
      produces:
      - application/json
      responses:
        "200":
          description: PaginatedResponse of Movies, total number of movies, or nil
            if none found
          schema:
            $ref: '#/definitions/pagination.PaginatedResponseOrdered-contracts_Movie'
        "400":
          description: Invalid request, invalid parameter or missing parameter
          schema:
            $ref: '#/definitions/apperrors.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apperrors.Error'
      summary: Get movies
      tags:
      - movies
    post:
      consumes:
      - application/json
      description: Create movie
      operationId: create-movie
      parameters:
      - description: Movie details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/contracts.CreateMovieRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Movie created
          schema:
            $ref: '#/definitions/contracts.MovieDetails'
        "400":
          description: Invalid request, invalid parameter or missing parameter
          schema:
            $ref: '#/definitions/apperrors.Error'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/apperrors.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apperrors.Error'
      summary: Create movie
      tags:
      - movies
  /movies/{movieId}:
    delete:
      description: Delete movie by id
      operationId: delete-movie-by-id
      parameters:
      - description: Movie ID
        in: path
        name: movieId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Movie deleted (softly deleting)
        "400":
          description: Invalid movie id, invalid parameter or missing parameter
          schema:
            $ref: '#/definitions/apperrors.Error'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/apperrors.Error'
        "404":
          description: Movie not found
          schema:
            $ref: '#/definitions/apperrors.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apperrors.Error'
      summary: Delete movie by id
      tags:
      - movies
    get:
      description: Get movie by id
      operationId: get-movie-by-id
      parameters:
      - description: Movie ID
        in: path
        name: movieId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Movie details
          schema:
            $ref: '#/definitions/contracts.MovieDetails'
        "400":
          description: Invalid movie id, invalid parameter or missing parameter
          schema:
            $ref: '#/definitions/apperrors.Error'
        "404":
          description: Movie not found
          schema:
            $ref: '#/definitions/apperrors.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apperrors.Error'
      summary: Get movie by id
      tags:
      - movies
    put:
      consumes:
      - application/json
      description: Update movie by id
      operationId: update-movie-by-id
      parameters:
      - description: Movie ID
        in: path
        name: movieId
        required: true
        type: integer
      - description: Updated movie details, fields that are not provided will not
          be updated
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/contracts.UpdateMovieRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Movie updated
          schema:
            $ref: '#/definitions/contracts.MovieDetails'
        "400":
          description: Invalid movie id, invalid parameter or missing parameter
          schema:
            $ref: '#/definitions/apperrors.Error'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/apperrors.Error'
        "404":
          description: Movie not found
          schema:
            $ref: '#/definitions/apperrors.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apperrors.Error'
      summary: Update movie by id
      tags:
      - movies
...
schemes:
- http
swagger: "2.0"